branch libft cuda // convertion double en float // mandelbrot done  //WIP debuf compil

+++ srcs/ft_julia_init_value.c
@@ -6,16 +6,16 @@
 /*   By: cledant <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/22 08:59:25 by cledant           #+#    #+#             */
-/*   Updated: 2016/01/22 09:58:43 by cledant          ###   ########.fr       */
+/*   Updated: 2016/07/28 21:29:17 by cledant          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
 
-double		ft_julia_init_value(int pos, size_t size, int axis)
+float		ft_julia_init_value(int pos, size_t size, int axis)
 {
 	if (axis == 0)
-		return (2 * (pos - ((int)size / 2)) / (double)(size / 2));
+		return (2 * (pos - ((int)size / 2)) / (float)(size / 2));
 	else
-		return (2 * (-pos + ((int)size / 2)) / (double)(size / 2));
+		return (2 * (-pos + ((int)size / 2)) / (float)(size / 2));
 }
